WiFi Robustness Applet - Refactor applet's menu structure
This change modifies the applet's menu structure so that each item is
implemented by way of a sub-menu (where it is not already).

The purpose is to present a better display of the help/explanatory text that
relates to each item. In particular to ensure that help text remains visible
while the user considers the relevant check box option.