Wifi robustness applet - Amend "Truncated Beacons" text
This change modifies the text display of the Truncated Beacons menu.

It highlights the possibility that the displayed count may be misleading
in circumstances where there is heavy system logging activity.

A failure of the WiFi subsystem may generate significant system logging.
Precisely the circumstances in which we may be examining the count...