Fix version reported for openvpn client.
Workaround a macro expansion change in v2.60 and fixed in v2.62 of
the auto configure tools.  Unfortunately bitbake is using v2.61 of
autoconf which exhibits the old expansion behaviour.  Previous attempts
to upgrade autoconf beyond 2.61 resulted in numerous package build
failures.

Details of the behaviour change as reported in the autoconf 2.62 changelog.

** The m4sugar macros m4_append and m4_append_uniq, first documented in
~   2.60, have been fixed to treat both the string and the separator
~   arguments consistently with regards to quoting.  Prior to this fix,
~   m4_append_uniq could mistakenly duplicate entries if the expansion
~   of the separator resulted in a different string (for example, if it
~   contained quotes, a comma, or a macro name).  However, it means
~   that programs previously using
~     m4_append([name], [string], [[, ]])
~   are now using a four-character separator instead of the intended
~   comma and space.